moved download demo data to uri.py, working on demo code
