+++ nems/utils.py
@@ -1,6 +1,7 @@
 import time
 import numpy as np
 
+
 def iso8601_datestring():
     '''
     Returns a string containing the present date as a string.
