+++ scripts/download_demo_data.py
@@ -1,9 +1,5 @@
-import logging
 import os
-import requests
-import nems.recording as recording
-
-log = logging.getLogger(__name__)
+import nems.uri
 
 # NOTE to LBHB team:
 #    If additional sample data is added in the future, just
@@ -12,71 +8,13 @@ log = logging.getLogger(__name__)
 #     be ferret1.tar.gz)
 #    Additionally, any sample data specified here should be in
 #    a zipped format, *not* a .mat, .csv etc.
-DEMO_NAMES = [
-        'TAR010c-18-1.tgz', 'eno052d-a1.tgz'
-        ]
-
-
-def get_demo_recordings(directory, unpack=False):
-    '''
-    Saves all sample recordings in the LBHB public s3 bucket to
-    nems/recordings/, or to the specified directory. By default,
-    the recordings will be kept in a compressed format; however,
-    specifying unpack=True will instead save them uncompressed
-    in a subdirectory.
-    '''
-    names = DEMO_NAMES
-    prefix = 'https://s3-us-west-2.amazonaws.com/nemspublic/sample_data/'
-    uris = [(prefix + n) for n in names]
-    if unpack:
-        recs = [recording.load(uri) for uri in uris]
-        for rec in recs:
-            log.info("Saving file at {} in {}".format(rec.uri, directory))
-            rec.save_dir(directory)
-    else:
-        """
-        https://stackoverflow.com/questions/16694907/
-        how-to-download-large-file-in-python-with-requests-py
-        """
-        for uri in uris:
-            file = uri.split('/')[-1]
-            local = os.path.join(directory, file)
-            if os.path.isfile(local):
-                log.info("Local file {} already exists, skipping.".format(local))
-            else:
-                log.info("Saving file at {} to {}".format(uri, local))
-                r = requests.get(uri, stream=True)
-                # TODO: clean this up, copied from recordings code.
-                #       All of these content-types have showed up *so far*
-                if not (r.status_code == 200 and
-                        (r.headers['content-type'] == 'application/gzip' or
-                         r.headers['content-type'] == 'application/x-gzip' or
-                         r.headers['content-type'] == 'application/x-compressed' or
-                         r.headers['content-type'] == 'application/x-compressed-tar' or
-                         r.headers['content-type'] == 'application/x-tar' or
-                         r.headers['content-type'] == 'application/x-tgz')):
-                    log.info('got response: {}, {}'
-                             .format(r.headers, r.status_code))
-                    raise Exception('Error loading from uri: {}'.format(uri))
-
-                try:
-                    with open(local, 'wb') as f:
-                        for chunk in r.iter_content(chunk_size=1024):
-                            if chunk:
-                                f.write(chunk)
-                except PermissionError as e:
-                    log.warn("Couldn't write in directory: \n{}\n"
-                             "due to permission issues. Make sure the "
-                             "parent directory grants write permission."
-                             .format(directory))
-                    log.exception(e)
 
 
 def main():
     import argparse
 
     relative_signals_dir = '/../recordings'
-    signals_dir = os.path.dirname(recording.__file__) + relative_signals_dir
+    signals_dir = os.path.dirname(nems.uri.__file__) + relative_signals_dir
     signals_dir = os.path.abspath(signals_dir)
 
     parser = argparse.ArgumentParser('Download NEMS demo data')
@@ -88,7 +26,7 @@ def main():
         help='Recordings compressed by default, set --unpack to decompress'
         )
     args = parser.parse_args()
-    get_demo_recordings(args.directory, args.unpack)
+    nems.uri.get_demo_recordings(args.directory, args.unpack)
 
 
 if __name__ == '__main__':
